Design: 디자인 시스템 팔레트 객체 추가
